website/pyproject.toml
[project]
name = "cross-auth-website"
version = "0.1.0"
description = "Cross-Auth documentation website"
requires-python = ">=3.10"
dependencies = [
    "cross-inertia",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "jinja2>=3.0.0",
]

[tool.uv]
dev-dependencies = []
